moved SessionInfo to session folder
