+++ client/app/session/session-info.ts
